Basic defending
Opponents are using the basic vertical stack tactic for defense.
The mark defense the throw side and the other defense opponents are marking the corresponding other (open) side.

Current problem:
Not very smooth/ideal, because the opponents are hitting the player. They are simply moving from one side to the other and not "around" the player.
They should away the player with some kind of path finding to get to their desired defense position.