#5 select player (#14)
* Players can be selected by clicking on them.

Selecting them:
* changes their color
* enables movement

All other players are unselected, i.e. they can't move anymore.
The Player holding the frisbee can not be selected.

* Possibility to cancel aiming

The aiming is canceled when the distance is not long enough.
This means if you aim and move the mouse back to the starting point, the frisbee won't be thrown.