Fixed skewed camera
After resuming to the play view (from the movement manager view) the camera angle was skewed.
This was because the stop condition to move the camera back was just depending on the position. When the camera reached the correct position, the rotation was not entirely finished resulting in a slightly skewed camera.

Fix: I could have added the rotation as an extra condition, but I decided to make the rotation faster instead. Now the camera is in the correct orientation, before the target position is reached.